% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptiveConfInts.R
\name{ciMean}
\alias{ciMean}
\title{ciMean}
\usage{
ciMean(x, conf.int = 0.95)
}
\arguments{
\item{x}{Vector of numeric data}

\item{conf.int}{confidence interval, between .01 and .99.}
}
\value{
Vector containing the lower and upper confidence interval
}
\description{
ciMean provides confidence intervals for a mean.
}
\examples{
#x <- rnorm(100)
#ciMean(x)
}

