% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrCompare.R
\name{corrCompare}
\alias{corrCompare}
\title{corrCompare}
\usage{
corrCompare(data = NULL, group = NULL, method = "pearson", rdec = 3,
  pdec = 3, tdec = 3)
}
\arguments{
\item{data}{data.frame containing 1 grouping variable and all variables to be used to compute a correlation table}

\item{group}{variable name of the grouping variable}

\item{method}{correlation method: pearson or spearman}

\item{rdec}{number of decimals for correlation coefficients}

\item{pdec}{number of decimals for p-values}

\item{tdec}{number of decimals for t-values}
}
\value{
A data.frame containing group level correlation coefficients as well as relevant tests between groups.
}
\description{
corrCompare computes a correlation table for each level of a binary grouping variable separately, and then tests differences in correlation coefficients between grouping variables
}
\examples{
#dat<-data.frame(g<-rep(c(1,2),50),
#                v2<-rnorm(100,5,1),
#                v1<-rnorm(100,2,1),
#                v3<-rnorm(100,499,12))
#names(dat) <- c("g","v1","v2","v3")
#jf <- corrCompare(data=dat, group="g")
#View(jf)
}

