% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmTable.R
\name{kmTable}
\alias{kmTable}
\title{kmTable}
\usage{
kmTable(data, time, event, group)
}
\arguments{
\item{data}{data.frame containing all data}

\item{time}{Variable name contaning time data}

\item{event}{Variable name containing event data (event = 1, censor = 0)}

\item{group}{Variable name containing groupiong variable}
}
\value{
R data.frame containing all km info
}
\description{
kmTable provides summary information from a kaplan-meier survival model, including median survival, event counts, and pairwise comparisons.
}
\examples{
#example_df <- data.frame(a = sample(c(0, 1), 100, replace = T),
#                         b = sample(letters[24:26], 100, replace = T),
#                         c = rnorm(100))
#kmTable(data=example_df, time = "c", event = "a", group = "b")
}

