% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multCox.R
\name{multCox}
\alias{multCox}
\title{multCox}
\usage{
multCox(data, timevar, event, vars, classvars = NULL, varLabelTable = NULL,
  dec = 2, pdec = 3)
}
\arguments{
\item{data}{Input data.frame}

\item{timevar}{Time variable}

\item{event}{Event variable, where event=1, censor=0}

\item{vars}{Vector of variable names to be modeled}

\item{classvars}{Vector of variable names that should be treated as categorical}

\item{varLabelTable}{A varLabelTable object that will convert variable names to nicer labels in the output table}

\item{dec}{Number of decimals for descriptive statistics}

\item{pdec}{Number of decimals for p-values}
}
\value{
Output table including the results from each univariate model
}
\description{
multCox will compute univariate Cox proportional hazards model models for multiple predictor variables.
}
\examples{
#NULL
}

