% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multLogistic.R
\name{multLogistic}
\alias{multLogistic}
\title{multLogistic}
\usage{
multLogistic(data, y, predlist, catlist = NULL, controlCatList = NULL,
  controlContinList = NULL, conditionalLogistic = FALSE, matchIDs = NULL,
  varLabelTable = NULL, minObsForML = 20, dec = 2, pdec = 3,
  rowwisePercent = T, includeCI = F, verbose = TRUE)
}
\arguments{
\item{data}{data.frame containing all data for analysis}

\item{y}{Outcome variable name}

\item{predlist}{Vector of predictor variable names}

\item{catlist}{Vector of categorical variable names}

\item{controlCatList}{Vector of categorical variable names to control for in each model}

\item{controlContinList}{Vetor of continuous variable names to control for in each model}

\item{conditionalLogistic}{If TRUE, perform conditional logisitic regression}

\item{matchIDs}{If conditionalLogistic==TRUE, variable name that contains matchIDs}

\item{varLabelTable}{varLabelTable object containing variable names in column 1 and variable labels in column 2}

\item{minObsForML}{The minimum number of events (y==1) observed in the dataset before the method should be switched to Firth's Penalized Maximum Likelihood. Default is 20, which is commonly used as a good rule of thumb.}

\item{dec}{Number of decimals to round descriptive statistics and odds ratios to}

\item{pdec}{Number of decimals to round p-values to}

\item{rowwisePercent}{If TRUE, present row-wise percents for descriptive statistics}

\item{includeCI}{If TRUE, confidence intervals for the means/proportions are included in the output.}

\item{verbose}{If TRUE, print additional information to the console}
}
\value{
An R data.frame object contianing all results.
}
\description{
multLogistic performs univariate, multivariate, and/or conditional logistic regression models over a list of predictor variables.
}
\examples{
#NULL
}

