% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multOutcomesLogistic.R
\name{multOutcomesLogistic}
\alias{multOutcomesLogistic}
\title{multOutcomesLogistic}
\usage{
multOutcomesLogistic(data, predictor, outcomes, predictorAsFactor = FALSE,
  rowwise = T, dec = 2, pdec = 3, verbose = TRUE)
}
\arguments{
\item{data}{data.frame containing all data for analyses}

\item{predictor}{Variable name for the predictor variable}

\item{outcomes}{Vector of variable names for the outcome variables}

\item{predictorAsFactor}{If TRUE, the predictor variable will be treated as a factor}

\item{rowwise}{If TRUE, row-wise percents are provided in the case of categorical predictors}

\item{dec}{Number of decimals to round descriptive}

\item{pdec}{Number of decimals to round p-vales}

\item{verbose}{Logical: should extra information be printed to the console}
}
\value{
A data.frame R object is returned
}
\description{
multOutcomesLogistic provides results for univariate binary logistic regression models for a single predictor and multiple outcomes.
}
\examples{
#NULL
}

