% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multReg.R
\name{multReg}
\alias{multReg}
\title{multReg}
\usage{
multReg(data, yvar, predlist, catlist = NULL, varLabelTable = FALSE,
  sortVars = FALSE, dec = 2, pdec = 3, rsqdec = 4)
}
\arguments{
\item{data}{data.frame containing all variables for analysis}

\item{yvar}{Variable name of the outcome variable}

\item{predlist}{Vector of predictor variables}

\item{catlist}{Vector of variable names found in `predlist` that should be treated as categorical variables}

\item{varLabelTable}{A varLabelTable object containing variable names in column 1 and variable labels in column 2}

\item{sortVars}{Vector of variable names in the order to be presented in the output}

\item{dec}{Number of decimals to round descriptive statistics}

\item{pdec}{Number of decimals to round p-values}

\item{rsqdec}{Number of decimals to round r-squared values}
}
\value{
A data.frame object conatining all results.
}
\description{
multReg performs univariate simple linear regression models for a single outcome variable over a list of predictor variables.
}
\examples{
#NULL
}

