% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskDifference.R
\name{riskDifference}
\alias{riskDifference}
\title{riskDifference}
\usage{
riskDifference(data, y, x)
}
\arguments{
\item{data}{data.frame containing x and y}

\item{y}{outcome variable name}

\item{x}{categorical predictor variable name}
}
\value{
A data.frame object containing relative risks with 95% CI
}
\description{
riskDifference provides the risk difference for a categorical predictor and binary outcome. The reference/control group is the first level of the predictor variable.
}
\examples{
#test_df <- data.frame(x=sample(c(1,2,3), 50, replace = T),
#                      y=sample(c(0,1), 50, replace = T))
#riskDifference(test_df, x = 'x', y = 'y')
}

