% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPercent.R
\name{survPercent}
\alias{survPercent}
\title{survPercent}
\usage{
survPercent(model, times, labels = NULL)
}
\arguments{
\item{model}{A survfit(Surv()) model from the survival package}

\item{times}{A vector of times to provide survival percentages for}

\item{labels}{Labels for the grouping variable, if applicable}
}
\value{
A data.frame containing survival percentages.
}
\description{
survPercent provides survival percentages for a survfit(Surv()) model for specified times.
}
\examples{
#example_dat <- data.frame(time = runif(100, 2, 100),
#                          event = sample(c(0, 1), 100, prob = c(.2, .8), replace = T),
#                          grp = sample(letters[1:3], 100, replace = T))
#library(survival)
#sMod <- survfit(Surv(time, event) ~ grp, data = example_dat)
#survPercent(model = sMod, times = c(10, 20, 30, 40, 50))
#survPercent(model = sMod, times = 1:5*10, labels=c("Group 1", "Group 2", "Group 3"))
}

