% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survTrunc.R
\name{survTrunc}
\alias{survTrunc}
\title{survTrunc}
\usage{
survTrunc(data, time, event, group, times, grpLabels = NULL,
  timeLabels = NULL)
}
\arguments{
\item{data}{data.frame containing all variables}

\item{time}{variable name for the time variable}

\item{event}{variable name for the event variable}

\item{group}{variable name for the group variable}

\item{times}{vector of times to truncate and test survival curves at}

\item{grpLabels}{labels for the group variable}

\item{timeLabels}{labels for the time variable}
}
\value{
data.frame containing results
}
\description{
survTrunc iteratively truncates patient observations at specified time intervals and provides Kaplan-Meier survival curves analysis at each.
}
\examples{
#NULL
}

