% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_testList.R
\name{t_testList}
\alias{t_testList}
\title{t_testList}
\usage{
t_testList(data, list_of_t.tests, dec = 2, pdec = 3, var.equal = TRUE,
  parametric = TRUE, paired = FALSE, NPDescriptives = F, NPtype = "IQR")
}
\arguments{
\item{data}{data.frame containing all columns provided in `list_of_t.tests`}

\item{list_of_t.tests}{A list, in which each element contains the names of two variables from `data`}

\item{dec}{Number of decimals to round descriptive statistics to}

\item{pdec}{Number of decimals to round p-values to}

\item{var.equal}{If TRUE, and parametric==TRUE, student's t.test will be used, else Welch's t.test methods are used}

\item{parametric}{If TRUE, t.test() methods are used, else wilcox.test() methods are used}

\item{paired}{If TRUE, the two variables will be treated as paired.}

\item{NPDescriptives}{If TRUE, Median and either IQR or range are provided, else Mean and Std are provided}

\item{NPtype}{Can be 'IQR' or 'range'. If NPDescriptives==TRUE, then the specified variance estimate(s) will be provided}
}
\value{
A data.frame containing the results from the t.tests along with descriptive statistics.
}
\description{
t_testList performs paired and unpaired parametric and non-parametric t.tests between columns within a data.frame, for multiple pairs of varaibles given in list form.
}
\examples{
#example_df <- data.frame(x1 = rnorm(100),
#                         x2 = rnorm(100),
#                         x3 = rnorm(100),
#                         x4 = rnorm(100),
#                         x5 = rnorm(100))
#list <- list(c("x1", "x2"), c("x1", "x3"), c("x4", "x5"))
#t_testList(example_df, list)
#t_testList(example_df, list, parametric = F)
#t_testList(example_df, list, paired = T)
#t_testList(example_df, list, NPDescriptives = T)
#t_testList(example_df, list, NPDescriptives = T, NPtype = "range")
}

