% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoWayTableWithMarginalPs.R
\name{twoWayTableWithMarginalPs}
\alias{twoWayTableWithMarginalPs}
\title{twoWayTableWithMarginalPs}
\usage{
twoWayTableWithMarginalPs(data, columnGroup, rowGroup, var,
  columnPaired = FALSE, rowPaired = FALSE, aovPairedID)
}
\arguments{
\item{data}{data.frame contianing all data}

\item{columnGroup}{Grouping variable to stratify across columns}

\item{rowGroup}{Grouping variable to stratify across rows}

\item{var}{Outcome variable, numeric}

\item{columnPaired}{Logical, whether or not the column grouping variable should be treated as paired}

\item{rowPaired}{Logical, whether or not the row grouping variable should be treated as paired}

\item{aovPairedID}{If paired, id variable name.}
}
\value{
data.frame with all output
}
\description{
twoWayTableWithMarginalPs
}
\examples{
# Needs and example
}

